Match wave twenty two packet and anim functions

+++ include/Battle/btlPacket.h
@@ -59,7 +59,7 @@ struct BtlPacket
 
 BtlPacket* btlPacketFindFirstByActionUID(u64 actionUID, u64 mask);
 u32 btlPacketCount();
-u32 btlPacketCountById(u16 id);
+u32 btlPacketCountById(u32 id);
 void btlPacket0027e4d0();
 BtlPacket* btlPacketCreate(u32 id, u32 workDataSize);
 u64 btlPacketRegister(BtlPacket* packet, u8 type);
